+++ lib/chirp-workflow/core/src/ctx/standalone.rs
@@ -4,14 +4,13 @@ use serde::Serialize;
 use uuid::Uuid;
 
 use crate::{
+	builder::common as builder,
 	ctx::{
-		api::WORKFLOW_TIMEOUT,
+		common,
 		message::{SubscriptionHandle, TailAnchor, TailAnchorResponse},
-		workflow::SUB_WORKFLOW_RETRY,
-		MessageCtx, OperationCtx,
+		MessageCtx,
 	},
 	db::DatabaseHandle,
-	error::WorkflowError,
 	error::WorkflowResult,
 	message::{Message, ReceivedMessage},
 	operation::{Operation, OperationInput},
@@ -70,158 +69,27 @@ impl StandaloneCtx {
 }
 
 impl StandaloneCtx {
-	pub async fn dispatch_workflow<I>(&self, input: I) -> GlobalResult<Uuid>
-	where
-		I: WorkflowInput,
-		<I as WorkflowInput>::Workflow: Workflow<Input = I>,
-	{
-		let name = I::Workflow::NAME;
-		let id = Uuid::new_v4();
-
-		tracing::info!(workflow_name=%name, workflow_id=%id, ?input, "dispatching workflow");
-
-		// Serialize input
-		let input_val = serde_json::to_value(input)
-			.map_err(WorkflowError::SerializeWorkflowOutput)
-			.map_err(GlobalError::raw)?;
-
-		self.db
-			.dispatch_workflow(self.ray_id, id, &name, None, input_val)
-			.await
-			.map_err(GlobalError::raw)?;
-
-		tracing::info!(workflow_name=%name, workflow_id=%id, "workflow dispatched");
-
-		Ok(id)
-	}
-
-	pub async fn dispatch_tagged_workflow<I>(
-		&self,
-		tags: &serde_json::Value,
-		input: I,
-	) -> GlobalResult<Uuid>
-	where
-		I: WorkflowInput,
-		<I as WorkflowInput>::Workflow: Workflow<Input = I>,
-	{
-		let name = I::Workflow::NAME;
-		let id = Uuid::new_v4();
-
-		tracing::info!(workflow_name=%name, workflow_id=%id, ?tags, ?input, "dispatching tagged workflow");
-
-		// Serialize input
-		let input_val = serde_json::to_value(input)
-			.map_err(WorkflowError::SerializeWorkflowOutput)
-			.map_err(GlobalError::raw)?;
-
-		self.db
-			.dispatch_workflow(self.ray_id, id, &name, Some(tags), input_val)
-			.await
-			.map_err(GlobalError::raw)?;
-
-		tracing::info!(workflow_name=%name, workflow_id=%id, "workflow dispatched");
-
-		Ok(id)
-	}
-
 	/// Wait for a given workflow to complete.
 	/// 60 second timeout.
 	pub async fn wait_for_workflow<W: Workflow>(
 		&self,
 		workflow_id: Uuid,
 	) -> GlobalResult<W::Output> {
-		tracing::info!(workflow_name=%W::NAME, id=?workflow_id, "waiting for workflow");
-
-		tokio::time::timeout(WORKFLOW_TIMEOUT, async move {
-			let mut interval = tokio::time::interval(SUB_WORKFLOW_RETRY);
-			loop {
-				interval.tick().await;
-
-				// Check if state finished
-				let workflow = self
-					.db
-					.get_workflow(workflow_id)
-					.await
-					.map_err(GlobalError::raw)?
-					.ok_or(WorkflowError::WorkflowNotFound)
-					.map_err(GlobalError::raw)?;
-				if let Some(output) = workflow.parse_output::<W>().map_err(GlobalError::raw)? {
-					return Ok(output);
-				}
-			}
-		})
-		.await?
+		common::wait_for_workflow::<W>(&self.db, workflow_id).await
 	}
 
-	/// Dispatch a new workflow and wait for it to complete. Has a 60s timeout.
-	pub async fn workflow<I>(
-		&self,
-		input: I,
-	) -> GlobalResult<<<I as WorkflowInput>::Workflow as Workflow>::Output>
+	/// Creates a workflow builder.
+	pub fn workflow<I>(&self, input: I) -> builder::workflow::WorkflowBuilder<I>
 	where
 		I: WorkflowInput,
 		<I as WorkflowInput>::Workflow: Workflow<Input = I>,
 	{
-		let workflow_id = self.dispatch_workflow(input).await?;
-		self.wait_for_workflow::<I::Workflow>(workflow_id).await
+		builder::workflow::WorkflowBuilder::new(self.db.clone(), self.ray_id, input)
 	}
 
-	/// Dispatch a new workflow with tags and wait for it to complete. Has a 60s timeout.
-	pub async fn tagged_workflow<I>(
-		&self,
-		tags: &serde_json::Value,
-		input: I,
-	) -> GlobalResult<<<I as WorkflowInput>::Workflow as Workflow>::Output>
-	where
-		I: WorkflowInput,
-		<I as WorkflowInput>::Workflow: Workflow<Input = I>,
-	{
-		let workflow_id = self.dispatch_tagged_workflow(tags, input).await?;
-		self.wait_for_workflow::<I::Workflow>(workflow_id).await
-	}
-
-	pub async fn signal<T: Signal + Serialize>(
-		&self,
-		workflow_id: Uuid,
-		input: T,
-	) -> GlobalResult<Uuid> {
-		let signal_id = Uuid::new_v4();
-
-		tracing::info!(signal_name=%T::NAME, %workflow_id, %signal_id, "dispatching signal");
-
-		// Serialize input
-		let input_val = serde_json::to_value(input)
-			.map_err(WorkflowError::SerializeSignalBody)
-			.map_err(GlobalError::raw)?;
-
-		self.db
-			.publish_signal(self.ray_id, workflow_id, signal_id, T::NAME, input_val)
-			.await
-			.map_err(GlobalError::raw)?;
-
-		Ok(signal_id)
-	}
-
-	pub async fn tagged_signal<T: Signal + Serialize>(
-		&self,
-		tags: &serde_json::Value,
-		input: T,
-	) -> GlobalResult<Uuid> {
-		let signal_id = Uuid::new_v4();
-
-		tracing::info!(signal_name=%T::NAME, ?tags, %signal_id, "dispatching tagged signal");
-
-		// Serialize input
-		let input_val = serde_json::to_value(input)
-			.map_err(WorkflowError::SerializeSignalBody)
-			.map_err(GlobalError::raw)?;
-
-		self.db
-			.publish_tagged_signal(self.ray_id, tags, signal_id, T::NAME, input_val)
-			.await
-			.map_err(GlobalError::raw)?;
-
-		Ok(signal_id)
+	/// Creates a signal builder.
+	pub fn signal<T: Signal + Serialize>(&self, body: T) -> builder::signal::SignalBuilder<T> {
+		builder::signal::SignalBuilder::new(self.db.clone(), self.ray_id, body)
 	}
 
 	#[tracing::instrument(err, skip_all, fields(operation = I::Operation::NAME))]
@@ -233,25 +101,15 @@ impl StandaloneCtx {
 		I: OperationInput,
 		<I as OperationInput>::Operation: Operation<Input = I>,
 	{
-		tracing::info!(?input, "operation call");
-
-		let ctx = OperationCtx::new(
-			self.db.clone(),
+		common::op(
+			&self.db,
 			&self.conn,
 			self.ray_id,
 			self.op_ctx.req_ts(),
 			false,
-			I::Operation::NAME,
-		);
-
-		let res = I::Operation::run(&ctx, &input)
-			.await
-			.map_err(WorkflowError::OperationFailure)
-			.map_err(GlobalError::raw);
-
-		tracing::info!(?res, "operation response");
-
-		res
+			input,
+		)
+		.await
 	}
 
 	pub async fn subscribe<M>(
